Reorganized code around in neat packages
