--- common.go
@@ -1,8 +1,0 @@
-package main
-
-var (
-	Buildstamp         string = "Unknown build timestamp"
-	Buildgithash       string = "Unknown git hash"
-	Buildhost          string = "Unknown build host"
-	Buildversionstring string = "0.0.1"
-)
