+++ common/common.go
@@ -1,0 +1,8 @@
+package common
+
+var (
+	Buildstamp         string = "Unknown build timestamp"
+	Buildgithash       string = "Unknown git hash"
+	Buildhost          string = "Unknown build host"
+	Buildversionstring string = "0.0.1"
+)
